Add new partition type API and convert GPT type checks.
This introduces a new type() function for partmaps modeled after the
label() and uuid() functions for filesystems. A likely future extension
will be support partition labels and uuids as well. This is in
preparation for adding more functionality like a search.part_type
command.

The msdos partition code is only partially converted for now, I'm not
sure about the preferred way to do this. Re-reading a msdos partition
table is not as easy as re-reading gpt is. One option would be to turn
'msdostype' into a union or data pointer that partmap modules can stash
extra info in. That would also allow gpt to avoid re-reading. Being new
to grub, are there memory usage concerns I should be aware of here?